Fixed sphere height movement
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@77 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/TSphere.h
@@ -29,7 +29,7 @@ class TSphere
         GLuint mID;
         static GLfloat mVData[12][3];
         static GLuint mTIndices[20][3];
-        Vector3D<int> mTranslation;
+        Vector3D<float> mTranslation;
         //Vector3D<double>** mRows;
 };
